+++ CHANGES.md
@@ -19,6 +19,7 @@ Release Notes.
 * Fix a bug in Spring Cloud Gateway if HttpClientFinalizer#send does not invoke, the span created at NettyRoutingFilterInterceptor can not stop.
 * Fix not tracing in HttpClient v5 when HttpHost(arg[0]) is null but `RoutingSupport#determineHost` works.
 * Support across thread tracing for SOFA-RPC.
+* Update Jedis 4.x plugin to support Sharding and Cluster models.
 
 #### Documentation
 * Update docs to describe `expired-plugins`.